--- api/src/main/java/com/personalwebsite/api/Message/MessageServiceImpl.java
@@ -1,5 +1,0 @@
-package com.personalwebsite.api.Message;
-
-public class MessageServiceImpl {
-    
-}
