--- api/src/main/java/com/personalwebsite/api/Role/Role.java
@@ -1,5 +1,0 @@
-package com.personalwebsite.api.Role;
-
-public class Role {
-    
-}
