--- api/src/main/java/com/personalwebsite/api/User/UserServiceImpl.java
@@ -1,5 +1,0 @@
-package com.personalwebsite.api.User;
-
-public class UserServiceImpl {
-    
-}
