+++ api/src/main/java/com/personalwebsite/api/config/SecurityConfig.java
@@ -1,4 +1,4 @@
-package com.personalwebsite.api.Security;
+package com.personalwebsite.api.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
