+++ api/src/main/java/com/personalwebsite/api/controller/PostController.java
@@ -1,4 +1,4 @@
-package com.personalwebsite.api.Post;
+package com.personalwebsite.api.controller;
 
 import org.springframework.web.bind.annotation.RestController;
