+++ api/src/main/java/com/personalwebsite/api/model/Post.java
@@ -1,4 +1,4 @@
-package com.personalwebsite.api.Post;
+package com.personalwebsite.api.model;
 
 import javax.persistence.Entity;
 import javax.persistence.Id;
