+++ api/src/main/java/com/personalwebsite/api/model/Role.java
@@ -1,0 +1,5 @@
+package com.personalwebsite.api.model;
+
+public class Role {
+    
+}
