+++ api/src/main/java/com/personalwebsite/api/model/User.java
@@ -1,34 +1,40 @@
-package com.personalwebsite.api.User;
+package com.personalwebsite.api.model;
 
-import java.util.HashSet;
-import java.util.Set;
+import java.util.Collection;
 
+import javax.persistence.ElementCollection;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import javax.persistence.Table;
 
 import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
 import lombok.Data;
 
 @Data
 @Entity
+@Table(name = "Users")
 public class User implements UserDetails {
     
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long id;
+
     private String username;
     private String password;
-    private Set<GrantedAuthority> authorities = new HashSet<GrantedAuthority>();
+    
+    @ElementCollection(targetClass = GrantedAuthority.class)
+    private Collection<? extends GrantedAuthority> authorities;
+
+    private boolean enabled = true;
 
     public User() {}
 
-    public User(String username, String password, 
-        Set<GrantedAuthority> authorities) {
-
+    public User(String username, String password, Collection<SimpleGrantedAuthority> authorities) {
         this.username = username;
         this.password = password;
         this.authorities = authorities;
@@ -36,21 +42,21 @@ public class User implements UserDetails {
 
     @Override
     public boolean isEnabled() {
-        return true;
+        return enabled;
     }
 
     @Override
     public boolean isCredentialsNonExpired() {
-        return true;
+        return enabled;
     }
 
     @Override
     public boolean isAccountNonExpired() {
-        return true;
+        return enabled;
     }
 
     @Override
     public boolean isAccountNonLocked() {
-        return true;
+        return enabled;
     }
 }
