+++ api/src/main/java/com/personalwebsite/api/repository/PostRepository.java
@@ -1,4 +1,6 @@
-package com.personalwebsite.api.Post;
+package com.personalwebsite.api.repository;
+
+import com.personalwebsite.api.model.Post;
 
 import org.springframework.data.jpa.repository.JpaRepository;
