+++ api/src/main/java/com/personalwebsite/api/service/messageservice/MessageService.java
@@ -1,0 +1,5 @@
+package com.personalwebsite.api.service.messageservice;
+
+public class MessageService {
+    
+}
