+++ api/src/main/java/com/personalwebsite/api/service/messageservice/MessageServiceImpl.java
@@ -1,0 +1,5 @@
+package com.personalwebsite.api.service.messageservice;
+
+public class MessageServiceImpl {
+    
+}
