+++ api/src/main/java/com/personalwebsite/api/service/userservice/UserServiceImpl.java
@@ -1,0 +1,5 @@
+package com.personalwebsite.api.service.userservice;
+
+public class UserServiceImpl {
+    
+}
